Refactor RPC implementation
