+++ node/Cargo.toml
@@ -58,8 +58,9 @@ frame-benchmarking-cli = { version = "4.0.0-dev", git = "https://github.com/pari
 
 # Local Dependencies
 rmrk-substrate-runtime = { version = "4.0.0-dev", path = "../runtime" }
-rmrk-rpc = { version = "0.0.1", path = "../rpc" }
-rmrk-traits = { version = "0.0.1", path = "../traits" }
+pallet-rmrk-rpc-runtime-api = { version = "0.0.1", path = "../rpc-runtime-api" }
+pallet-rmrk-rpc = { version = "0.0.1", path = "../rpc" }
+rmrk-traits = { version = "0.0.1", default-features = false, path = "../traits" }
 pallet-rmrk-core = { version = "0.0.1", path = "../pallets/rmrk-core" }
 pallet-rmrk-equip = { version = "0.0.1", path = "../pallets/rmrk-equip" }
